git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_bvoc@5936 70805675-57b6-444b-bb00-18ecffcb4be8

+++ modules/spinupdata.cpp
@@ -70,7 +70,7 @@ void GenericSpinupData::detrend_data() {
 	}
 
 	double a, b;
-	regress(&year_number.front(), &annual_mean.front(), nbr_years(), a, b);
+	regress(&year_number.front(), &annual_mean.front(), (int) nbr_years(), a, b);
 
 	for (size_t y = 0; y < nbr_years(); ++y) {
 		double anomaly = b*y;
